requirements.txt
Django==1.2.1
Pillow==2.3.0
South==0.7.1
django-extensions==0.5
django-pagination==1.0.6

psycopg2
pytils==0.2.3

#sorl-thumbnail==11.12
git+https://github.com/xyz666/sorl-thumbnail.git@dev#egg=sorl

#Djapian==2.3.1
https://github.com/xyz666/djapian-2.3.1-patched/archive/master.zip

#django-cms==2.1.0.beta3
-e git://github.com/divio/django-cms.git@56120ab0416998207b2a0ca0318f9b3ecc9d56b8#egg=django-cms

xlrd
xlwt
suds==0.4
python-memcached==1.53
django-tinymce==1.5.3
django-endless-pagination==2.0
lxml==3.4.4
django-cbv==0.2

django-recaptcha==1.3.0
supercaptcha==0.1.4

gunicorn